requirements.txt
aiosmtpd==1.2
atpublic==1.0
